1.51

+++ macro/ocrpy.py
@@ -4,6 +4,7 @@ import pyautogui as pag
 import numpy as np
 from logpy import log
 import os
+import subprocess
 import time
 def loadRes():
     outdict =  {}
@@ -25,8 +26,23 @@ def millify(n):
     n = float(n)
     millidx = max(0,min(len(millnames)-1,
                         int(math.floor(0 if n == 0 else math.log10(abs(n))/3))))
-
+def screenshot(**kwargs):
+    out = None
+    for _ in range(4):
+        try: 
+            if "region" in kwargs:
+                out = pag.screenshot(region=kwargs['region'])
+            else:
+                out = pag.screenshot()
+            break
+        except FileNotFoundError as e:
+            log(e)
+            print(e)
+            time.sleep(0.5)
+    return out
+    
 def imToString(m):
+    setdat = loadsettings.load()
     savedata = loadRes()
     ww = savedata['ww']
     wh = savedata['wh']
@@ -35,40 +51,44 @@ def imToString(m):
     ylm = loadsettings.load('multipliers.txt')['y_length_multiplier']
     xlm = loadsettings.load('multipliers.txt')['x_length_multiplier']
     sn = time.time()
-    # Path of tesseract executable
-    #pytesseract.pytesseract.tesseract_cmd ='**Path to tesseract executable**'
-    # ImageGrab-To capture the screen image in a loop. 
-    # Bbox used to capture a specific area.
+    ebY = wh//(20*ysm)
+    honeyY = 0
+    if setdat["new_ui"]:
+        ebY = wh//(14*ysm)
+        honeyY = 31
+        if setdat['display_type'] == "built-in retina display": honeyY*=2
     if m == "bee bear":
-        cap = pag.screenshot(region=(ww//(3*xsm),wh//(20*ysm),ww//(3*xlm),wh//(7*ylm)))
+        cap = screenshot(region=(ww//(2.7*xsm),ebY/1.1,ww//(3*xlm),wh//(15*ylm)))
     elif m == "egg shop":
-        cap = pag.screenshot(region=(ww//(1.2*xsm),wh//(3*ysm),ww-ww//1.2,wh//5))
+        cap = screenshot(region=(ww//(1.2*xsm),wh//(3*ysm),ww-ww//1.2,wh//5))
     elif m == "blue":
-        cap = pag.screenshot(region=(ww*3//4, wh//3*2, ww//4,wh//3))
+        cap = screenshot(region=(ww*3//4, wh//3*2, ww//4,wh//3))
     elif m == "chat":
-        cap = pag.screenshot(region=(ww*3//4, 0, ww//4,wh//3))
+        cap = screenshot(region=(ww*3//4, 0, ww//4,wh//3))
     elif m == "ebutton":
-        cap = pag.screenshot(region=(ww//(2.65*xsm),wh//(20*ysm),ww//(21*xlm),wh//(17*ylm)))
-        cap.save("{}.png".format(sn))  
+        cap = screenshot(region=(ww//(2.65*xsm),ebY,ww//(21*xlm),wh//(17*ylm)))
+        if not cap: return ""
+        cap.save("{}.png".format(sn))
         result = ocr.ocr("{}.png".format(sn),cls=False)[0]
-        result = sorted(result, key = lambda x: x[1][1], reverse = True)
         os.remove("{}.png".format(sn))
         try:
+            result = sorted(result, key = lambda x: x[1][1], reverse = True)
             return result[0][1][0]
         except:
             return ""
     elif m == "honey":
-        cap = pag.screenshot(region=(ww//(3*xsm),0,ww//(6.5*xlm),wh//(ylm*25)))
+        cap = pag.screenshot(region=(ww//(3*xsm),honeyY,ww//(6.5*xlm),wh//(ylm*25)))
+        if not cap: return ""
         cap.save("{}.png".format(sn))  
         ocrres = ocr.ocr("{}.png".format(sn),cls=False)[0]
-        #print(ocrres)
-        result = [x[1][0] for x in ocrres]
         honey = 0
-        for i in result:
-            if i[0].isdigit():
-                honey = i
-                break
+        #print(ocrres)
         try:
+            result = [x[1][0] for x in ocrres]
+            for i in result:
+                if i[0].isdigit():
+                    honey = i
+                    break
             honey = int(''.join([x for x in honey if x.isdigit()]))
             log(millify(honey))
         except Exception as e:
@@ -76,13 +96,17 @@ def imToString(m):
         os.remove("{}.png".format(sn))
         return honey
     elif m == "disconnect":
-        cap = pag.screenshot(region=(ww//(3),wh//(2.8),ww//(2.3),wh//(5)))
+        cap = screenshot(region=(ww//(3),wh//(2.8),ww//(2.3),wh//(5)))
     elif m == "dialog":
-        cap = pag.screenshot(region=(ww//(3*xsm),wh//(1.6*ysm),ww//(8*xlm),wh//(ylm*15)))
+        cap = screenshot(region=(ww//(3*xsm),wh//(1.6*ysm),ww//(8*xlm),wh//(ylm*15)))
+    if not cap: return ""
     cap.save("{}.png".format(sn))  
     result = ocr.ocr("{}.png".format(sn),cls=False)[0]
-    result = sorted(result, key = lambda x: x[1][1], reverse = True)
-    out = ''.join([x[1][0] for x in result])
+    try:
+        result = sorted(result, key = lambda x: x[1][1], reverse = True)
+        out = ''.join([x[1][0] for x in result])
+    except:
+        out = ""
     os.remove("{}.png".format(sn))
     log("OCR for {}\n\n{}".format(m,out))
     return out
@@ -93,11 +117,14 @@ def customOCR(X1,Y1,W1,H1,applym=1):
     ylm = loadsettings.load('multipliers.txt')['y_length_multiplier']
     xlm = loadsettings.load('multipliers.txt')['x_length_multiplier']
     if applym:
-        cap = pag.screenshot(region=(X1/xsm,Y1/ysm,W1/xlm,H1/ylm))
+        cap = screenshot(region=(X1/xsm,Y1/ysm,W1/xlm,H1/ylm))
     else:
-        cap = pag.screenshot(region=(X1,Y1,W1,H1))
+        cap = screenshot(region=(X1,Y1,W1,H1))
     cap.save("{}.png".format(sn)) 
     out = ocr.ocr("{}.png".format(sn),cls=False)
     log("OCR for Custom\n{}".format(out))
     os.remove("{}.png".format(sn))
-    return out[0]
+    if not out is None:
+        return out[0]
+    else:
+        return [[[""],["",0]]]
